requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
webdriver-manager==4.0.1
deep-translator==1.11.4
reportlab==4.0.7
PyPDF2==3.0.1
python-telegram-bot==20.7
python-dotenv==1.0.0
pytz==2023.3
pytest==8.3.5
